// src/types/Etccd.d.ts
export interface etccd {
    etccd_id: string;
    parent_id?: string;
    etccd_cate?: string;
    etccd_subcate?: string;
    etccd_value_1?: string;
    etccd_value_2?: string;
    etccd_value_3?: string;
    etccd_value_4?: string;
    etccd_value_5?: string;
    etccd_value_6?: string;
    etccd_useyn: boolean;
    add_user_id?: string;
    add_date?: Date;
    update_user_id?: string;
    update_date?: Date;
    etccd_img?: string;
}